Se envian datos menores a 512MB, correctamente. Existen fugas de memoria, ya que el codigo asume que lo gestiona el manejadorCache

+++ TP2SERVIDOR/ARCHIVO.h
@@ -5,12 +5,13 @@
 
 #ifndef _ARCHIVO_H
 #define _ARCHIVO_H
+#include <sys/types.h>
 using namespace std;
 
 struct ARCHIVO
 {
     char* puntero;
-    int size;
+    size_t size;
 };
 
 #endif //_ARCHIVO_H
